# apidocs/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: MediaServer API
  description: API of MediaServer, a server and proxy that supports various protocols.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9997
security: []
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    PathConfList:
      type: object
      properties:
        pageCount:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/PathConf'
...
    PathList:
      type: object
      properties:
        pageCount:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Path'
...
    HLSMuxer:
      type: object
      properties:
        path:
          type: string
        created:
          type: string
        lastRequest:
          type: string
        bytesSent:
          type: integer
          format: int64
    HLSMuxerList:
      type: object
      properties:
        pageCount:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/HLSMuxer'
...
    RTSPConn:
      type: object
      properties:
        id:
          type: string
        created:
          type: string
        remoteAddr:
          type: string
        bytesReceived:
          type: integer
          format: int64
        bytesSent:
          type: integer
          format: int64
    RTSPConnList:
      type: object
      properties:
        pageCount:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/RTSPConn'
    RTSPSession:
      type: object
      properties:
        id:
          type: string
        created:
          type: string
        remoteAddr:
          type: string
        state:
          type: string
          enum: [idle, read, publish]
        path:
          type: string
        transport:
          type: string
          nullable: true
        bytesReceived:
          type: integer
          format: int64
        bytesSent:
          type: integer
          format: int64
    RTSPSessionList:
      type: object
      properties:
        pageCount:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/RTSPSession'
...
paths:
  /v3/config/global/get:
    get:
      operationId: configGlobalGet
      summary: returns the global configuration.
      description: ''
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalConf'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/config/global/patch:
    patch:
      operationId: configGlobalSet
      summary: patches the global configuration.
      description: all fields are optional.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalConf'
      responses:
        '200':
          description: the request was successful.
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/config/pathdefaults/get:
    get:
      operationId: configPathDefaultsGet
      summary: returns the default path configuration.
      description: ''
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PathConf'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/config/pathdefaults/patch:
    patch:
      operationId: configPathDefaultsPatch
      summary: patches the default path configuration.
      description: all fields are optional.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PathConf'
      responses:
        '200':
          description: the request was successful.
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/config/paths/list:
    get:
      operationId: configPathsList
      summary: returns all path configurations.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PathConfList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v3/hlsmuxers/list:
    get:
      operationId: hlsMuxersList
      summary: returns all HLS muxers.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HLSMuxerList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v3/paths/list:
    get:
      operationId: pathsList
      summary: returns all paths.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PathList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v3/rtspconns/list:
    get:
      operationId: rtspConnsList
      summary: returns all RTSP connections.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPConnList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspconns/get/{id}:
    get:
      operationId: rtspConnsGet
      summary: returns a RTSP connection.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the connection.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPConn'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspsessions/list:
    get:
      operationId: rtspSessionsList
      summary: returns all RTSP sessions.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPSessionList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspsessions/get/{id}:
    get:
      operationId: rtspSessionsGet
      summary: returns a RTSP session.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the connection.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPSession'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspsessions/kick/{id}:
    post:
      operationId: rtspSessionsKick
      summary: kicks out a RTSP session from the server.
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ID of the session.
        schema:
          type: string
      responses:
        '200':
          description: the request was successful.
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v3/rtspsconns/list:
    get:
      operationId: rtspsConnsList
      summary: returns all RTSPS connections.
      description: ''
      parameters:
      - name: page
        in: query
        description: page number.
        schema:
          type: integer
          default: 0
      - name: itemsPerPage
        in: query
        description: items per page.
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: the request was successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RTSPConnList'
        '400':
          description: invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...